refactor: move type definitions from types.ts into domain/models.ts
